backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
spotipy==2.23.0
python-jose==3.3.0
flask-limiter==3.5.0
SQLAlchemy==2.0.21
gunicorn==21.2.0
pytest==7.4.2
black==23.9.1
flake8==6.1.0

